e2e tests refactoring: query endpoint tests
