+++ tests/e2e/__init__.py
@@ -1,1 +1,3 @@
 """End to end tests for the OLS service."""
+
+__all__ = ["test_api", "test_query_endpoint"]
